+++ src/product/model/productModel.go
@@ -14,7 +14,7 @@ type Product struct {
 	Size      productEnum.Size          `gorm:"not null" json:"size"`
 	Value     float32                   `gorm:"not null" json:"value"`
 	Fabric    string                    `gorm:"not null" json:"fabric"`
-	Stock     int64                     `gorm:"not null" json:"stock"`
+	Stock     uint32                     `gorm:"not null" json:"stock"`
 	CreatedAt time.Time                 `json:"created_at"`
 	UpdatedAt time.Time                 `json:"updated_at"`
 }
